[particleflow] clean up driftchamber hits merger
